MVC implemented
