+++ client/src/model/classConstants.ts
