+++ client/src/services/interface.ts
@@ -1,7 +1,6 @@
 export interface ICrud<T> {
   create(obj: T): void;
-  edit(obj: T): void;
+  read(): any;
   delete(obj: T): void;
   save(obj: T): void;
-  cancel(obj: T): void;
 }
